feat: remove MirrorableEnum
