+++ p2p_wallet/Extensions/String+Extensions.swift
@@ -202,3 +202,20 @@ extension String {
         prefix(1).uppercased() + dropFirst()
     }
 }
+
+extension String {
+    var snakeAndFirstUppercased: String? {
+        guard let snakeCase = snakeCased() else { return nil }
+        return snakeCase.prefix(1).uppercased() + snakeCase.dropFirst()
+    }
+    
+    func snakeCased() -> String? {
+        let pattern = "([a-z0-9])([A-Z])"
+
+        let regex = try? NSRegularExpression(pattern: pattern, options: [])
+        let range = NSRange(location: 0, length: count)
+        return regex?.stringByReplacingMatches(in: self, options: [], range: range, withTemplate: "$1_$2")
+            .uppercaseFirst
+    }
+}
+
