+++ p2p_wallet/Services/AnalyticsManager/Parameters/KeyAppAnalyticsParameter+Extensions.swift
@@ -1,16 +1,29 @@
 import Foundation
 import AnalyticsManager
 
-extension KeyAppAnalyticsParameter: MirrorableEnum {
+extension KeyAppAnalyticsParameter {
     var name: String? {
         mirror.label.snakeAndFirstUppercased
     }
 
     var value: Any? {
-        mirror.params.values.first
+        mirror.value
     }
     
     var providerIds: [AnalyticsProviderId] {
         [KeyAppAnalyticsProviderId.amplitude].map(\.rawValue)
+    }
+    
+    // MARK: - Helpers
+
+    private var mirror: (label: String, value: Any?) {
+        let reflection = Mirror(reflecting: self)
+        guard reflection.displayStyle == .enum,
+              let associated = reflection.children.first,
+              let label = associated.label
+        else {
+            return ("\(self)", nil)
+        }
+        return (label, associated.value)
     }
 }
